mq2: example of splitting the mq2 driver and fti components
The following is the log message (just the currently relevant parts)
from the original branch where I tried this:

    This commit separates the Function Table Interface (FTI) implementation
    into separate .h and .c files (mq2_fti.h, mq2_fti.c).

    This has several benefits:

    - since the FTI is essentially just a wrapper around the C driver,
      treat it as such -- similar to how the C++ drivers are implemented.

    - in the C driver implementation, we can use an actual driver context
      typedef rather than void* everywhere.  This helps promote
      type-safety during development, and makes the implementation of the
      driver entry points cleaner and easier to use.  For example:

      the driver mq2__init() function is now defined like:

      mq2_context mq2_init(int16_t pin);

      And this type is used to pass the context to all driver
      implementation functions.  So for example mq2_get_value() is
      defined:

      upm_result_t mq2_get_value(const mq2_context dev, float *value);

      And in that function we can reference the context directly like so:

      // old way
      *value = counts * ((upm_mq2*)dev)->m_count_scale;

      // new way
      *value = counts * dev->m_count_scale;

      since now the context is accessible directly rather than having to
      be casted through a void pointer all the time, which is a little ugly
      and potentially error prone.

      The code is much cleaner and easier to understand as a result.

    - In the driver implementation, we drop the 'upm_' in every function
      name.  There is no need for this, and it just makes the code
      confusing to read.

    - however, in the FTI implementation functions, we preserve the "upm_"
      prefix for all FTI functionality since it makes absolute sense there.

      In addition, this makes it pretty clear what belongs to FTI and what
      belongs to the driver.

      These functions would have to be named differently from the driver
      functions anyway to avoid symbol collisions, so keeping the "upm_"
      prefix in the FTI implementation neatly solves that problem as well.

    - In the _fti.[hc] implementation files, you can see that all the
      machinery for supporting FTI is present, and in the _fti.c file, you
      can see that the actual entry points are essentially just wrappers
      around the driver functions.

    - in the future, an option can be added to cmake so that the FTI
      implementation files are not even compiled.

      In reality - FTI will never be needed for straight-up C driver or
      C++/SWIG driver users anyway.  It will just confuse them.
