+++ src/mq2/CMakeLists.txt
@@ -1,6 +1,6 @@
 set (libname "mq2")
 set (libdescription "upm x_sensorname sensor")
-set (module_src ${libname}.cxx mq2.c)
+set (module_src ${libname}.cxx mq2.c mq2_fti.c)
 set (module_hpp ${libname}.hpp)
-set (module_h ${libname}.hpp mq2.h)
+set (module_h ${libname}.hpp mq2.h mq2_fti.h)
 upm_module_init()
