+++ src/mq2/mq2.hpp
@@ -129,7 +129,7 @@ namespace upm {
     float counts();
 
   private:
-    void* _dev;
+    mq2_context _dev;
   };
 }
